[#89914214] Renaming stock erp classes so they can be used with RFID stock.
