--- java/messages/src/main/java/com/nedap/retail/messages/erp/v1/ErpStockSummary.java
@@ -1,6 +1,0 @@
-package com.nedap.retail.messages.erp.v1;
-
-public class ErpStockSummary {
-
-    public String id;
-}
