+++ java/messages/src/main/java/com/nedap/retail/messages/stock/StockSummaryListRequest.java
@@ -1,4 +1,4 @@
-package com.nedap.retail.messages.erp.v1;
+package com.nedap.retail.messages.stock;
 
 import org.codehaus.jackson.map.annotate.JsonSerialize;
 import org.joda.time.DateTime;
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonInclude(Include.NON_NULL)
 @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
-public class ErpStockSummaryList {
+public class StockSummaryListRequest {
 
     public static final String FROM_EVENT_TIME = "from_event_time";
     public static final String UNTIL_EVENT_TIME = "until_event_time";
@@ -24,11 +24,11 @@ public class ErpStockSummaryList {
     @org.codehaus.jackson.annotate.JsonProperty(UNTIL_EVENT_TIME)
     public DateTime untilEventTime;
 
-    public ErpStockSummaryList() {
+    public StockSummaryListRequest() {
 
     }
 
-    public ErpStockSummaryList(final String location, final DateTime fromEventTime, final DateTime untilEventTime) {
+    public StockSummaryListRequest(final String location, final DateTime fromEventTime, final DateTime untilEventTime) {
         this.location = location;
         this.fromEventTime = fromEventTime;
         this.untilEventTime = untilEventTime;
