+++ java/messages/src/main/java/com/nedap/retail/messages/stock/StockSummaryRequest.java
@@ -1,0 +1,6 @@
+package com.nedap.retail.messages.stock;
+
+public class StockSummaryRequest {
+
+    public String id;
+}
